backend/requirement.txt
ibm-watson
python-dotenv
ibm-watsonx-ai
streamlit
pandas
PyMuPDF
matplotlib
plotly.express
reportlab
